requirements.txt
requests==2.26.0
kafka-python==2.0.2
mysql-connector-python==8.0.28
aiokafka==0.10.0
aiohttp==3.9.3
pytest==8.0.2
aioresponses==0.7.6
pytest-asyncio==0.23.5